fix(tests): disambiguate Color(uint8_t, int, ...) calls in cube/volume tests
Texture3DTests and TextureCubeTests each built a Color from a mix of
std::uint8_t and plain int arguments, which is ambiguous between
Color(intcs, intcs, intcs, intcs) and Color(bytecs, bytecs, bytecs, bytecs) --
neither overload is a better match, so the call fails to compile. Found while
building CnaTests for the HTML_DOM backend under emcc 6.0.5.

Casts the remaining int literals to std::uint8_t so both calls reach the bytecs
overload their sibling arguments already targeted. No behavioural change: the
values were already in byte range.

+++ tests/Microsoft/Xna/Framework/Graphics/Texture3DTests.cpp
@@ -195,7 +195,7 @@ TEST_F(Texture3DTest, SetDataNonZeroStartIndexUploadsOnlyItsOwnWindow)
     std::vector<Color> src(15, poison);
     for (int i = 0; i < 8; ++i)
         src[static_cast<std::size_t>(4 + i)] =
-            Color(static_cast<std::uint8_t>(20 + i * 11), 130,
+            Color(static_cast<std::uint8_t>(20 + i * 11), static_cast<std::uint8_t>(130),
                   static_cast<std::uint8_t>(200 - i * 9), static_cast<std::uint8_t>(150 + i));
     EXPECT_NO_THROW(tex.SetData(src.data(), 4, 8));
